refactor: move NotFoundPage component to ErrorPages folder
