+++ src/pages/ErrorPages/tests/NotFoundPage.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import '@testing-library/jest-dom';
 import { fireEvent, screen } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
-import NotFoundPage from './NotFoundPage';
+import NotFoundPage from '../NotFoundPage';
 import renderPage from '@/utils/tests/utilts';
 
 test('Should navigate to home page', async () => {
